requirements.txt
Django==2.0.0
gunicorn==19.7.1
wheel==0.33.4
whitenoise==4.1.3
djangorestframework==3.6.4
django-cors-headers==3.0.2
django-rest-auth==3.10.2

